telnyx_restaurant/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
jinja2>=3.1.0
python-dotenv>=1.0.0
